gtest/requirements.txt
aws-requests-auth==0.3.0
awsebcli==3.8.3
beautifulsoup4==4.5.1
boto==2.43.0
boto3==1.4.1
botocore==1.4.65
bs4==0.0.1
cement==2.8.2
colorama==0.3.7
Django==1.10.2
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.12
elasticsearch==5.0.0
futures==3.0.5
jmespath==0.9.0
oauthlib==2.0.0
pathspec==0.3.4
pysolr==3.6.0
python-dateutil==2.5.3
python-google-places==1.2.0
PyYAML==3.12
requests==2.12.1
requests-aws4auth==0.9
requests-oauthlib==0.7.0
s3transfer==0.1.9
semantic-version==2.5.0
six==1.10.0
texttable==0.8.6
tweepy==3.5.0
urllib3==1.18
virtualenv==15.0.3
watson-developer-cloud==0.22.0
websocket-client==0.37.0
